~ Updates preprocessor defines and the copyright lines. ~ Reduces number of warnings at compile stage.

+++ vgd/trunk/src/event/ButtonStateSet.cpp
@@ -1,4 +1,4 @@
-// VGSDK - Copyright (C) 2004, IRCAD.
+// VGSDK - Copyright (C) 2004-2006, Nicolas Papier.
 // Distributed under the terms of the GNU Library General Public License (LGPL)
 // as published by the Free Software Foundation.
 // Author Nicolas Papier
@@ -15,6 +15,8 @@ namespace vgd
 	
 namespace event
 {
+
+
 
 void ButtonStateSet::apply( const vgd::event::KeyboardButtonEvent *pKeyboardButtonEvent )
 {
@@ -109,7 +111,7 @@ void ButtonStateSet::setUp()
 
 uint32 ButtonStateSet::getNumDown() const
 {
-	return ( m_down.size() );
+	return ( static_cast<uint32>(m_down.size()) );
 }
